requirements.txt
praw
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
email-validator
dnspython
